Add pubsub wakeup for disconnected subscribers.
Currently this feature is always on (see NodeSubscription and NodeAffiliate);
however, it should be an option to the subscriber.  The 2nd phase is to change
the schema and client API to support the option.  The current implementation
uses a new WakeupProvider interface to remove the circular dependency between
Openfire and MMX plugin.